Updated mapgen and caddy deployment
In tucsonmesh/trello-to-geojson#5, the functionality for creating the
map geojson was changed to be provided via a Python package + CLI tool.

This commit adjusts the nixos configuration accordingly to build and
install this package on the path of the associated systemd service unit.

Not a huge fan of all the permission issues between mapgen and caddy but
we'll figure something out. Probably move the website to not be deployed
under /var/lib/mapgen long term. Symlink to the geojson generated by the
mapgen unit